django-polls/requirements.txt
Django==4.1.3
gunicorn==20.1.0
psycopg2-binary==2.9.5
python-dotenv==0.21.0
whitenoise==5.3.0
django-minio-storage==0.3.10
